+++ scripts/dev-bundle-tool-package.js
@@ -15,7 +15,7 @@ var packageJson = {
     pacote: "https://github.com/meteor/pacote/tarball/dea6a20a68057237a498f6c047e9b3ecf5df749d",
     "node-gyp": "3.6.2",
     "node-pre-gyp": "0.6.36",
-    "meteor-babel": "7.0.0-beta.3-1",
+    "meteor-babel": "7.0.0-beta.3-2",
     "meteor-promise": "0.8.6",
     promise: "8.0.1",
     reify: "0.12.3",